# contract/openapi.yaml
openapi: 3.1.0
info:
  title: cmProjectX Service API
  version: 0.1.0
  description: >
    Contract between the Rust/WinUI 3 client and the .NET Intune engine.
    This is the single source of truth for DTOs and endpoints — generate
    C# and Rust types from it (see contract/README.md).
servers:
  - url: http://127.0.0.1:5099
    description: Local sidecar (MVP)
paths:
...
  /apps:
    get:
      operationId: listApps
      summary: Intune applications (live Graph; 409 when signed out)
      responses:
        "200":
          description: Applications for the active tenant
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/AppListItem" }
        "409": { description: Not signed in }

  # ── Devices (M2): live Graph list endpoints, all returning ListItem[] ──
...
  /apps/{id}:
    get:
      operationId: getApp
      summary: Application object JSON (read-only)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
  /apps/{id}/assignments:
    get:
      operationId: getAppAssignments
      summary: Normalized assignment set for an application (group names resolved best-effort)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses:
        "200":
          description: The application's current assignments
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/Assignment" }
        "409": { description: Not signed in }
    post:
      operationId: setAppAssignments
      summary: Replace ALL assignments on an application (replace-all write)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: array
              items: { $ref: "#/components/schemas/Assignment" }
      responses:
        "204": { description: Assignments replaced }
        "409": { description: Not signed in }
  /apps/assign:
    post:
      operationId: bulkAssignApps
      summary: Apply the same assignments to many apps at once (replace-all per app; dryRun default true)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/BulkAssignRequest" }
      responses:
        "200":
          description: Per-app assign results
          content:
            application/json:
              schema: { $ref: "#/components/schemas/BulkAssignResult" }
        "400": { description: appIds is required }
        "409": { description: Not signed in }
  /app-protection:
    get:
      operationId: listAppProtection
      summary: App protection policies
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createAppProtection
      summary: Create an app protection policy
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /app-protection/{id}:
    get:
      operationId: getAppProtection
      summary: App protection policy object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateAppProtection
      summary: Update an app protection policy
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteAppProtection
      summary: Delete an app protection policy
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /app-configs:
    get:
      operationId: listAppConfigs
      summary: Managed-device app configurations
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
    post:
      operationId: createAppConfig
      summary: Create an app configuration
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "200": { $ref: "#/components/responses/Created" }, "409": { description: Not signed in } }
  /app-configs/{id}:
    get:
      operationId: getAppConfig
      summary: App configuration object JSON
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }
    patch:
      operationId: updateAppConfig
      summary: Update an app configuration
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      requestBody: { $ref: "#/components/requestBodies/ObjectJson" }
      responses: { "204": { description: Updated }, "409": { description: Not signed in } }
    delete:
      operationId: deleteAppConfig
      summary: Delete an app configuration
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "204": { description: Deleted }, "409": { description: Not signed in } }
  /vpp-tokens:
    get:
      operationId: listVppTokens
      summary: Apple VPP tokens (read-only)
      responses: { "200": { $ref: "#/components/responses/ListItemArray" }, "409": { description: Not signed in } }
  /vpp-tokens/{id}:
    get:
      operationId: getVppToken
      summary: VPP token object JSON (read-only)
      parameters: [ { name: id, in: path, required: true, schema: { type: string } } ]
      responses: { "200": { $ref: "#/components/responses/ObjectJson" }, "404": { description: Not found }, "409": { description: Not signed in } }

  # ── Enrollment ──
...
components:
  responses:
    ListItemArray:
      description: Normalized list rows for the active tenant
      content:
        application/json:
          schema:
            type: array
            items: { $ref: "#/components/schemas/ListItem" }
    ObjectJson:
      description: The full resource as a free-form Graph object (normalized JSON)
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
    Created:
      description: Created; returns the new object's id
      content:
        application/json:
          schema:
            type: object
            properties: { id: { type: string, nullable: true } }
  requestBodies:
    ObjectJson:
      description: The resource as a free-form Graph object (normalized JSON); the @odata.type drives deserialization
      required: true
      content:
        application/json:
          schema: { type: object, additionalProperties: true }
  schemas:
...
    AppListItem:
      type: object
      description: One Intune application (Graph MobileApp), flattened for the Applications list.
      properties:
        id:                   { type: string }
        displayName:          { type: string }
        description:          { type: string, nullable: true }
        publisher:            { type: string, nullable: true }
        appType:              { type: string, description: "Friendly app type, e.g. Win32, iOS Store" }
        platform:             { type: string, enum: [Windows, iOS, Android, macOS, Web, Other] }
        createdDateTime:      { type: string }
        lastModifiedDateTime: { type: string }
        isAssigned:           { type: boolean }
        publishingState:      { type: string }
        isFeatured:           { type: boolean }
      required: [id, displayName, appType, platform, createdDateTime, lastModifiedDateTime, isAssigned, publishingState, isFeatured]
    ListItem:
      type: object
      description: >
        Normalized list row used by the generic LIVE list screens (Devices, and
        later Identity/Admin). Per-feature endpoints project their Graph element
        type into this shape server-side.
      properties:
        id:       { type: string }
        title:    { type: string }
        subtitle: { type: string }
        badge:    { type: string, nullable: true }
        platform: { type: string, nullable: true }
        modified: { type: string, nullable: true }
        source:   { type: string, nullable: true, description: "Backing MDM provider (e.g. 'jamf') when the row is from a non-Intune provider (M22); null = Intune." }
      required: [id, title, subtitle]
...
    Assignment:
      type: object
      description: >
        Normalized Intune assignment row shared by every assignable surface
        (mirrors api_types::Assignment / AssignmentDto byte-for-byte). The sidecar
        translates this to/from the per-resource typed Graph assignment target.
        `intent` is apps-only (null elsewhere).
      properties:
        kind:       { type: string, description: "group | exclusionGroup | allDevices | allUsers" }
        groupId:    { type: string, nullable: true }
        groupName:  { type: string, nullable: true }
        filterId:   { type: string, nullable: true }
        filterMode: { type: string, nullable: true, description: "include | exclude" }
        intent:     { type: string, nullable: true, description: "apps: required | available | uninstall | availableWithoutEnrollment" }
      required: [kind]
    BulkAssignRequest:
      type: object
      description: POST /apps/assign body — the same assignments applied to every app in appIds.
      properties:
        appIds:      { type: array, items: { type: string } }
        assignments: { type: array, items: { $ref: "#/components/schemas/Assignment" } }
        dryRun:      { type: boolean, nullable: true, description: "default true — preview without writing" }
      required: [appIds, assignments]
    BulkAssignItemResult:
      type: object
      properties:
        appId: { type: string }
        ok:    { type: boolean }
        error: { type: string, nullable: true }
      required: [appId, ok]
    BulkAssignResult:
      type: object
      properties:
        results: { type: array, items: { $ref: "#/components/schemas/BulkAssignItemResult" } }
      required: [results]
...